+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Discussions about unicorn-binance-websocket-api releases!](https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/discussions/categories/releases)
 
 ## 1.42.0.dev (development stage/unreleased/unstable)
+### Added
+- Logging to `manager.wait_till_stream_has_started()`
+- Logging to `manager.wait_till_stream_has_stoped()`
+### Fixed
+- Better "stop" measuring in `manager.wait_till_stream_has_stoped()`
+
 
 ## 1.42.0
 ### Added